# openapi.yaml
openapi: 3.0.3
info:
  title: Appliance Shop API - OpenAPI 3.0
  description: |-
    This is a Home Appliances Store Server based on the OpenAPI 3.0 specification. API for device management in Home Appliance Stores.

    Some useful links:
    - [The Appliance Shop repository](https://github.com/minhcopilot/Appliance-Shop-API)
    - [The source API definition for the Appliance Shop](https://github.com/minhcopilot/Appliance-Shop-API/openopenapi.yaml)

  contact:
    email: phanminh190502@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:9000
    description: Local server
...
paths:
...
  /orders:
    get:
      tags:
        - order
      security:
        - bearerAuth: []
      summary: Get all orders
      responses:
        200:
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        204:
          description: No orders found
        500:
          description: Internal server error
    post:
      tags:
        - order
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        500:
          description: Internal server error
  /orders/momo-payment:
    post:
      tags:
        - order
      summary: Create a new order and initiate MoMo payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        200:
          description: Payment initiation response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 'Payment initiated'
        400:
          description: Bad request
        500:
          description: Internal server error
  /orders/callback:
    post:
      tags:
        - order
      summary: Handle MoMo payment callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resultCode:
                  type: integer
                  example: 0
                orderId:
                  type: string
                  example: '123456789'
      responses:
        200:
          description: Order status updated successfully
        400:
          description: Payment not successful or order not found
        500:
          description: Internal server error
  /orders/transaction-status:
    post:
      tags:
        - order
      summary: Get transaction status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                  example: '123456789'
      responses:
        200:
          description: Transaction status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'COMPLETED'
  /orders/search:
    get:
      tags:
        - order
      summary: Search orders based on a keyword
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          required: true
          description: Keyword to search orders
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Keyword is required
        500:
          description: Internal server error
  /orders/customer-orders:
    get:
      tags:
        - order
      summary: Get orders by customer email or phone number
      parameters:
        - in: query
          name: email
          schema:
            type: string
          required: false
        - in: query
          name: phoneNumber
          schema:
            type: string
          required: false
      responses:
        200:
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Email or phone number is required
        500:
          description: Internal server error
  /orders/customer/{customerId}:
    get:
      tags:
        - order
      summary: Get orders by customer ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        404:
          description: Customer not found
        500:
          description: Internal server error
  /orders/{id}:
    get:
      tags:
        - order
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        403:
          description: Forbidden
        404:
          description: Order not found
        500:
          description: Internal server error
  /orders/user/{userId}:
    get:
      tags:
        - order
      summary: Get orders by user ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        403:
          description: Forbidden
        404:
          description: Customer not found
        500:
          description: Internal server error
  /orders/{orderId}:
    patch:
      tags:
        - order
      summary: Update an order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        200:
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          description: Order not found
        500:
          description: Internal server error
...
components:
  schemas:
...
    OrderDetail:
      type: object
      properties:
        productId:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        price:
          type: number
          example: 100.0
        discount:
          type: number
          example: 10
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        shippedDate:
          type: string
          format: date-time
          example: '2023-06-10T15:00:00Z'
        status:
          type: string
          example: 'PENDING'
        description:
          type: string
          example: 'Order description'
        shippingAddress:
          type: string
          example: '123 Street, City'
        shippingCity:
          type: string
          example: 'City'
        paymentType:
          type: string
          example: 'CREDIT_CARD'
        customerId:
          type: integer
          example: 1
        employeeId:
          type: integer
          example: 2
        orderDetails:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetail'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT